requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
redis>=5.0.0
celery[redis]>=5.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
firebase-admin>=6.4.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pandas>=2.2.0
openpyxl>=3.1.0
boto3>=1.34.0
openai>=1.3.0
reportlab>=4.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
python-dotenv>=1.0.0 